+++ Monitor/src/main/java/racecontrol/client/extension/statistics/CarProperties.java
@@ -75,6 +75,11 @@ public interface CarProperties {
     public final Key<Integer> RACE_START_POSITION = new Key<>(Integer.class, 0);
     public final Key<Boolean> RACE_START_POSITION_ACCURATE = new Key<>(Boolean.class, false);
     public final Key<Integer> PLACES_GAINED = new Key<>(Integer.class, 0);
+    
+    public final Key<Integer> PITLANE_TIME = new Key<>(Integer.class, 0);
+    public final Key<Integer> PITLANE_TIME_STATIONARY = new Key<>(Integer.class, 0);
+    public final Key<Integer> PITLANE_COUNT = new Key<>(Integer.class, 0);
+    public final Key<Boolean> PITLANE_COUNT_ACCURATE = new Key<>(Boolean.class, false);
 
     // Session
     public final Key<Integer> SESSION_BEST_LAP_TIME = new Key<>(Integer.class, 0);