+++ Monitor/src/main/java/racecontrol/client/extension/statistics/StatisticsExtension.java
@@ -15,6 +15,7 @@ import racecontrol.client.events.CarConnectedEvent;
 import racecontrol.client.extension.statistics.processors.DataProcessor;
 import racecontrol.client.extension.statistics.processors.GapProcessor;
 import racecontrol.client.extension.statistics.processors.OvertakeProcessor;
+import racecontrol.client.extension.statistics.processors.PitTimeProcessor;
 import racecontrol.client.extension.statistics.processors.PlacesLostGainedProcessor;
 import racecontrol.client.extension.statistics.processors.RealtimePositionProcessor;
 import racecontrol.client.extension.statistics.processors.SectorTimesProcessor;
@@ -66,6 +67,7 @@ public class StatisticsExtension
         processors.add(new GapProcessor(cars));
         processors.add(new OvertakeProcessor(cars));
         processors.add(new PlacesLostGainedProcessor(cars));
+        processors.add(new PitTimeProcessor(cars));
     }
 
     @Override